google-gemini/flash-2-0/requirements.txt
Pillow>=10.0.0
python-dotenv>=1.0.0
google-genai>=1.11.0